a21s-common: Kang OMX stack from gta4xlwifi
* This allows us to remove our somewhat "falty"
* shim, and using the the gta4xlwifi stack built
* from BSP seems to fix the following errors.
* (these errors previously broke video recording)

* E Exynos_OSAL_Android: [lockBuffer] Failed to mapper.lock()
* E Exynos_OSAL_Android: [Exynos_OSAL_LockMetaData]: Failed to lockBuffer (err:0x80001001)
* E EXYNOS_VIDEO_ENCCONTROL: [Exynos_Shared_BufferToData]: Failed to Exynos_OSAL_LockMetaData (err:0x80001001)

+++ shims/libshim_ui/GraphicBufferMapper.cpp
@@ -23,7 +23,7 @@ status_t _ZN7android19GraphicBufferMapper9lockAsyncEPK13native_handlemmRKNS_4Rec
 }
 
 status_t _ZN7android19GraphicBufferMapper9lockAsyncEPK13native_handleyyRKNS_4RectEPPviPiS9_(
-        void* thisptr, buffer_handle_t handle, uint64_t producerUsage, uint64_t consumerUsage,
+        void* thisptr, buffer_handle_t handle, uint32_t producerUsage, uint32_t consumerUsage,
         const Rect& bounds, void** vaddr, int fenceFd, int32_t* /*outBytesPerPixel*/,
         int32_t* /*outBytesPerStride*/) {
     auto* gpm = static_cast<android::GraphicBufferMapper*>(thisptr);
